Revert "Revert ItemRenderer removal (#847)"
This reverts commit 45350a2458cd3f816bc87af40997e426feb4b13b.
